pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = "A program for obtaining weather for a specified area"
authors = ["Alikhan Mulaev <alikhan.earth.love@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
